refactor: extract range and order traits
